+++ src/main/java/com/barlinc/unusual_prehistory/blocks/plant/update_6/NeedleLitterBlock.java
@@ -1,0 +1,10 @@
+package com.barlinc.unusual_prehistory.blocks.plant.update_6;
+
+import net.minecraft.world.level.block.SnowyDirtBlock;
+
+public class NeedleLitterBlock extends SnowyDirtBlock {
+
+    public NeedleLitterBlock(Properties properties) {
+        super(properties);
+    }
+}