+++ src/main/java/com/barlinc/unusual_prehistory/registry/UP2Trees.java
@@ -6,6 +6,7 @@ import com.barlinc.unusual_prehistory.worldgen.feature.tree.ginkgo.GinkgoFoliage
 import com.barlinc.unusual_prehistory.worldgen.feature.tree.lepidodendron.HangingLepidodendronLeavesDecorator;
 import com.barlinc.unusual_prehistory.worldgen.feature.tree.lepidodendron.LepidodendronFoliagePlacer;
 import com.barlinc.unusual_prehistory.worldgen.feature.tree.lepidodendron.LepidodendronTrunkPlacer;
+import com.barlinc.unusual_prehistory.worldgen.feature.tree.metasequoia.MetasequoiaFoliagePlacer;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.world.level.levelgen.feature.foliageplacers.FoliagePlacerType;
 import net.minecraft.world.level.levelgen.feature.treedecorators.TreeDecoratorType;
@@ -25,6 +26,7 @@ public class UP2Trees {
     // Foliage placers
     public static final DeferredHolder<FoliagePlacerType<?>, FoliagePlacerType<GinkgoFoliagePlacer>> GINKGO_FOLIAGE = FOLIAGE_PLACERS.register("ginkgo_foliage_placer", () -> new FoliagePlacerType<>(GinkgoFoliagePlacer.CODEC));
     public static final DeferredHolder<FoliagePlacerType<?>, FoliagePlacerType<LepidodendronFoliagePlacer>> LEPIDODENDRON_FOLIAGE = FOLIAGE_PLACERS.register("lepidodendron_foliage_placer", () -> new FoliagePlacerType<>(LepidodendronFoliagePlacer.CODEC));
+    public static final DeferredHolder<FoliagePlacerType<?>, FoliagePlacerType<MetasequoiaFoliagePlacer>> METASEQUOIA_FOLIAGE = FOLIAGE_PLACERS.register("metasequoia_foliage_placer", () -> new FoliagePlacerType<>(MetasequoiaFoliagePlacer.CODEC));
 
     // Tree decorators
     public static final DeferredHolder<TreeDecoratorType<?>, TreeDecoratorType<?>> HANGING_LEPIDODENDRON_LEAVES = TREE_DECORATORS.register("hanging_lepidodendron_leaves", () -> new TreeDecoratorType<>(HangingLepidodendronLeavesDecorator.CODEC));